# lumen/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mario Baggio test",
        "contact": {
            "name": "Mario Baggio",
            "email": "mario.baggio@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/adcampaigns/{name}/{page_number}": {
            "get": {
                "tags": [
                    "AdCampaigns"
                ],
                "description": "AdCampaigns list",
                "operationId": "App\\Http\\Controllers\\AdcampaignsController::getAdcampaigns",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "AdCampaign name",
                        "required": false,
                        "allowEmptyValue": true,
                        "example": "Xmas"
                    },
                    {
                        "name": "page_number",
                        "in": "path",
                        "description": "Page number",
                        "required": false,
                        "allowEmptyValue": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "AdCampaigns list"
                    }
                }
            }
        },
        "/adcampaigns/{name}/{cash_back_rate}/{date_from}/{date_to}": {
            "post": {
                "tags": [
                    "AdCampaigns"
                ],
                "description": "New AdCampaign",
                "operationId": "App\\Http\\Controllers\\AdcampaignsController::create",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "AdCampaign name",
                        "required": true,
                        "example": "Xmas in July"
                    },
                    {
                        "name": "cash_back_rate",
                        "in": "path",
                        "description": "AdCampaign cash_back_rate (in %)",
                        "required": true,
                        "example": "5.5"
                    },
                    {
                        "name": "date_from",
                        "in": "path",
                        "description": "AdCampaign Start date",
                        "required": true,
                        "example": "2020-07-01 15:00:00"
                    },
                    {
                        "name": "date_to",
                        "in": "path",
                        "description": "AdCampaign End date",
                        "required": true,
                        "example": "2020-07-03 15:00:00"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "New AdCampaign addedd"
                    },
                    "412": {
                        "description": "Precondition Failed"
                    }
                }
            }
        },
        "/adcampaigns/{id}/{new_name}/{new_cash_back_rate}/{new_date_from}/{new_date_to}": {
            "patch": {
                "tags": [
                    "AdCampaigns"
                ],
                "description": "Update AdCampaign",
                "operationId": "App\\Http\\Controllers\\AdcampaignsController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "AdCampaign ID",
                        "required": true,
                        "example": "1"
                    },
                    {
                        "name": "new_name",
                        "in": "path",
                        "description": "AdCampaign new name",
                        "required": false,
                        "example": "Xmas in December"
                    },
                    {
                        "name": "new_cash_back_rate",
                        "in": "path",
                        "description": "Adcampaign new cash_back_rate",
                        "required": false,
                        "example": "10.2"
                    },
                    {
                        "name": "new_date_from",
                        "in": "path",
                        "description": "Adcampaign new date_from",
                        "required": false,
                        "example": "2020-02-02"
                    },
                    {
                        "name": "new_date_to",
                        "in": "path",
                        "description": "Adcampaign new date_to",
                        "required": false,
                        "example": "2020-02-20"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Adcampaign updated"
                    },
                    "412": {
                        "description": "Precondition Failed"
                    }
                }
            }
        },
        "/adcampaigns/{id}": {
            "delete": {
                "tags": [
                    "AdCampaigns"
                ],
                "description": "Delete Adcampaign",
                "operationId": "App\\Http\\Controllers\\AdcampaignsController::delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Adcampaign id",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Adcampaign deleted"
                    },
                    "412": {
                        "description": "Precondition Failed"
                    }
                }
            }
        },
        "/adcampaigns/{adcampaign_id}/merchants/{page_number}": {
            "get": {
                "tags": [
                    "AdCampaigns-Merchants"
                ],
                "description": "AdCampaigns merchants list",
                "operationId": "App\\Http\\Controllers\\AdcampaignsController::getAdcampaignMerchants",
                "parameters": [
                    {
                        "name": "adcampaign_id",
                        "in": "path",
                        "description": "AdCampaign ID",
                        "required": true,
                        "example": "1"
                    },
                    {
                        "name": "page_number",
                        "in": "path",
                        "description": "Page number",
                        "required": false,
                        "allowEmptyValue": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "AdCampaigns merchants list"
                    }
                }
            }
        },
        "/adcampaigns/{adcampaign_id}/merchants/{merchant_id}": {
            "post": {
                "tags": [
                    "AdCampaigns-Merchants"
                ],
                "description": "New merchant in AdCampaign",
                "operationId": "App\\Http\\Controllers\\AdcampaignsController::createAdcampaignMerchants",
                "parameters": [
                    {
                        "name": "adcampaign_id",
                        "in": "path",
                        "description": "AdCampaign id",
                        "required": true,
                        "example": "1"
                    },
                    {
                        "name": "merchant_id",
                        "in": "path",
                        "description": "Merchant id",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Merchant addedd"
                    },
                    "412": {
                        "description": "Precondition Failed"
                    }
                }
            },
            "delete": {
                "tags": [
                    "AdCampaigns-Merchants"
                ],
                "description": "Remove Merchant from Adcampaign",
                "operationId": "App\\Http\\Controllers\\AdcampaignsController::deleteAdcampaignMerchants",
                "parameters": [
                    {
                        "name": "adcampaign_id",
                        "in": "path",
                        "description": "AdCampaign id",
                        "required": true,
                        "example": "1"
                    },
                    {
                        "name": "merchant_id",
                        "in": "path",
                        "description": "Merchant id",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Adcampaign deleted"
                    },
                    "412": {
                        "description": "Precondition Failed"
                    }
                }
            }
        },
...
        "/merchants/{id}": {
            "delete": {
                "tags": [
                    "Merchants"
                ],
                "description": "Delete Merchant",
                "operationId": "App\\Http\\Controllers\\MerchantsController::delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Merchant id",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Merchant deleted"
                    },
                    "412": {
                        "description": "Precondition Failed"
                    }
                }
            }
        },
...
        "/sitemap_categories/{id}": {
            "delete": {
                "tags": [
                    "Sitemap categories"
                ],
                "description": "Delete Sitemap Category",
                "operationId": "App\\Http\\Controllers\\SitemapCategoriesController::delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sitemap cateogory deleted"
                    },
                    "412": {
                        "description": "Precondition Failed"
                    }
                }
            }
        }
    }
}